+++ backend/controllers/DefaultController.php
@@ -1,16 +1,24 @@
 <?php
+/**
+ * @link https://github.com/black-lamp/yii2-newsletter
+ * @copyright Copyright (c) Vladimir Kuprienko
+ * @license BSD 3-Clause License
+ */
+
 namespace bl\newsletter\backend\controllers;
 
-use yii;
-use yii\data\Pagination;
+use Yii;
+use yii\base\Exception;
+use yii\data\ActiveDataProvider;
 use yii\web\Controller;
 use yii\web\Response;
 
+use bl\newsletter\backend\Module as Newsletter;
 use bl\newsletter\common\entities\Client;
 use bl\newsletter\common\helpers\CSV;
 
 /**
- * Default controller for the `newsletter` module
+ * Default controller for the backend module
  *
  * @author Vladimir Kuprienko <vldmr.kuprienko@gmail.com>
  */
@@ -20,6 +28,11 @@ class DefaultController extends Controller
      * @inheritdoc
      */
     public $defaultAction = 'list';
+    /**
+     * @var Newsletter
+     */
+    public $module;
+
 
     /**
      * List of subscribed clients
@@ -28,34 +41,26 @@ class DefaultController extends Controller
      */
     public function actionList()
     {
-        $query = Client::find();
+        $providerConfig = array_merge($this->module->dataProvider, ['query' => Client::find()]);
+        $provider = new ActiveDataProvider($providerConfig);
 
-        $count_query = clone $query;
-        $pages = new Pagination([
-            'totalCount' => $count_query->count()
-        ]);
-
-        $clients = $query->offset($pages->offset)
-            ->limit($pages->limit)
-            ->all();
-
-        $view_params = [
-            'clients' => $clients,
-            'pages' => $pages
+        $viewParams = [
+            'provider' => $provider,
+            'enableCsv' => $this->module->enableCsv
         ];
 
         if(Yii::$app->request->isAjax) {
-            return $this->renderAjax('list', $view_params);
+            return $this->renderAjax('list', $viewParams);
         }
 
-        return $this->render('list', $view_params);
+        return $this->render('list', $viewParams);
     }
 
     /**
      * Remove client from database
      *
      * @param integer $id
-     * @return Response
+     * @return Response|string
      */
     public function actionDelete($id)
     {
@@ -64,7 +69,33 @@ class DefaultController extends Controller
         }
 
         if(Yii::$app->request->isAjax) {
-            $this->actionList();
+            return $this->actionList();
+        }
+
+        return $this->redirect(Yii::$app->request->referrer);
+    }
+
+    /**
+     * Remove all clients from database
+     *
+     * @return string|Response
+     */
+    public function actionClear()
+    {
+        $clients = Client::find()->all();
+        $transaction = Client::getDb()->beginTransaction();
+        try {
+            foreach ($clients as $client) {
+                $client->delete();
+            }
+            $transaction->commit();
+        }
+        catch (Exception $ex) {
+            $transaction->rollBack();
+        }
+
+        if(Yii::$app->request->isAjax) {
+            return $this->actionList();
         }
 
         return $this->redirect(Yii::$app->request->referrer);
@@ -75,7 +106,6 @@ class DefaultController extends Controller
      */
     public function actionDownloadCsv()
     {
-        $csv = Client::getCsv();
-        CSV::download($csv, 'clients.csv');
+        CSV::download($this->module->clientManager->getCsv(), 'clients.csv');
     }
 }